remove example.rs code. Ensure the fragments API takes the fragment size without conversion

+++ src/packet/serialized.rs
@@ -20,8 +20,8 @@ impl SerializedPacket {
     }
 
     /// Returns an iterator yielding payload fragments
-    pub fn fragments(&self, fragment_size: usize) -> impl Iterator<Item = &[u8]> {
+    pub fn fragments(&self, fragment_size: u16) -> impl Iterator<Item = &[u8]> {
         /// TODO: This needs to be implemented to yield out fragments based on size
-        self.payload.chunks(fragment_size)
+        self.payload.chunks(fragment_size as usize)
     }
 }
